[GpuCommandBuffer] Move creation of ScopedGpuGraphExec into subclasses
The `ScopedGpuGraphExec` type is an RAII wrapper that changes the executable graph handle of a nested command buffer to a given value and restores it to its old value during destruction.

It's needed to handle updates of an existing executable graph. Since nested command buffers don't have their own executable graph this wrapper helps with making the executable graph of the root command buffer available to its nested command buffers.

This change makes all of this happen in platform specific sublasses by providing an pure virtual function declaration `ActivateScopedGpuGraphExec` which returns a unique pointer to a base class of the RAII wrapper.

PiperOrigin-RevId: 688081283