[XLA] Speed up the HLO DCE pass.
As in HloModule::RemoveUnusedComputations, we can simply perform a DFS over the computations both to find used computations and to trigger RunOnComputation on any used computations. We then remove all unreachable computations.

This improves over the existing code because we avoid repeatedly calling HloModule::MakeComputationPostOrder(), each call to which requires traversing every instruction in every computation. Instead, we can perform a single traversal of the computations.

PiperOrigin-RevId: 689859898